refactor: take the archived/active choice onto the base
`Choice` moves out of `ui/controls.tsx` into `ui/choice.tsx`, rebuilt on
Tailwind. The roving tab index, the arrow, Home and End handling, the
radiogroup semantics and every prop are unchanged; three consumers change
one import line.

It stays a radio group rather than becoming shadcn's Tabs or ToggleGroup.
Both would draw this and both would say something else about it: tabs name
panels a page switches between, and this switches which rows the table is
asked for. What a person's assistive technology is told does not change
because the styling did.

The press feedback keeps the shape the rest of the base uses —
`transition-transform` on the named property, `scale-97` only on
`:active:not(:focus-visible)` so a keyboard activation is immediate, and a
`motion-reduce` form that removes both.
