+++ apps/web/app/projects/[projectId]/tests/page.tsx
@@ -20,7 +20,7 @@ import {
   type ListedTest,
   type TestPage,
 } from "../../../../lib/tests.ts";
-import { Choice } from "../../../../ui/controls.tsx";
+import { Choice } from "../../../../ui/choice.tsx";
 import { Toolbar } from "../../../../ui/section.tsx";
 import { DataTable, type Column } from "../../../../ui/data-table.tsx";
 import { Empty, Failure, Loading, NotFound } from "../../../../ui/page-state.tsx";
